Distinguish between N_UNDF and unknown
This distinction is important because of the special case where 'value' is nonzero, 'type' is N_UNDF, and the symbol does not exist in any of the other binaries being handled by the linker. In this case, the linker is responsible for allocating space in .bss for this symbol, with its size in bytes equal to the 'value' field here.